requirements.txt
python-dotenv~=0.17.1
discord.py~=1.7.2
peewee~=3.14.4